Better compatibility with default DOOM modeline
Improves legibility in the modeline for both the project directory and the file path in a modified buffer when using DOOM Emacs. The text colour will be the same as the rest of the foreground text in the modeline, which is the behaviour in Vanilla Emacs, and the text weighting will change in a modified buffer to signify unsaved changes. The changes do not affect the default behaviour of the theme on Vanilla Emacs.